requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Advanced ML (optional)
xgboost>=1.7.0
lightgbm>=3.3.0

# Backtesting and metrics
scipy>=1.9.0

# Configuration
pyyaml>=6.0

# HTTP requests
requests>=2.28.0

# Model persistence
joblib>=1.2.0

# Development and testing
pytest>=7.2.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=6.0.0
mypy>=0.990

# Logging and monitoring
python-dotenv>=0.21.0

# Jupyter (for notebook compatibility)
jupyter>=1.0.0
ipykernel>=6.19.0
